Event Data (clean) added
This file contains event data for every Jadon Sancho action in the 20/21 CL season.
Data contains starting and ending (where applicable) coordinates for every action.

A big part of the process to prepare the data for analysis was to assign location (left, right, or center) to each row.  We used a 100x100 pitch in this project where anything from 0 to 30 (inclusive) on the Y-axis was deemed right. Any action between 30 and 70 was assigned center, and any event greater than or equal to 70 was assigned the value left.

The data cleaning and prepping process also involved confirming that the values in the outcome column were correct as there were mistakes in the raw data.